Refacto components into a commun hook
